Changes in data collecting module
- renaming `Data Loader` to `Data Collector`
- added logging and retrying
- added missing db engine `psycopg2` to the requirements
