+++ data-usage-monitor.ini.example
@@ -6,4 +6,32 @@ month_statistics_endpoint=api/monitoring/month_statistics
 host=http://127.0.0.1
 flask_port=5000
 flask_endpoint=api/usage-stamp
-data_loader_job_schedule=10
+data_collector_job_schedule=10
+
+[loggers]
+keys=root,dataCollector
+
+[handlers]
+keys=consoleHandler
+
+[formatters]
+keys=simpleFormatter
+
+[logger_root]
+level=DEBUG
+handlers=consoleHandler
+
+[logger_dataCollector]
+level=DEBUG
+handlers=consoleHandler
+qualname=dataCollector
+propagate=0
+
+[handler_consoleHandler]
+class=StreamHandler
+level=DEBUG
+formatter=simpleFormatter
+args=(sys.stdout,)
+
+[formatter_simpleFormatter]
+format=%(asctime)s - %(name)s - %(levelname)s - %(message)s
