+++ requirements.txt
@@ -4,6 +4,7 @@ lxml==4.8.0
 requests==2.27.1
 schedule==1.1.0
 python-dotenv==0.20.0
+psycopg2==2.9.3
 SQLAlchemy==1.4.35
 Flask-SQLAlchemy==2.5.1
 Flask-Migrate==3.1.0
